refactor(job details): replace the routing jobId params with hashing
- rename page from jobs to job for consistency
- adapt links in the Jobs List
- miscellaneous fixing and config
