+++ next.config.ts
@@ -1,7 +1,7 @@
 import type {NextConfig} from "next";
 
 const nextConfig: NextConfig = {
-    // output: "export",
+    output: "export",
     sassOptions: {
         silenceDeprecations: ["legacy-js-api"],
     }
