+++ src/components/Applications/Forms/InsertEditJob/props.types.ts
@@ -1,3 +1,5 @@
+import {JobEntry} from "@/types/JobEntry";
+
 export default interface Props {
-    data?: null | "";
+    data?: null | JobEntry;
 }
